Move VarInt template literal type to top of file structure for reuse
