+++ 2-Computer-Systems/problems/1-Protobuf-Varint/src/varint.d.ts
@@ -1,0 +1,1 @@
+export type VarInt = `0x${string}`;
